punto funzionante della flexo con combinazione cilindri e altezze bobine.

+++ PapiroMVC/Models/PlotterRoll.cs
@@ -20,6 +20,6 @@ namespace PapiroMVC.Models
     {
     	
     	[DataMember] 		
-        public Nullable<long> Width { get; set; }
+        public Nullable<double> Width { get; set; }
     }
 }
